+++ app/src/main/java/com/example/databases/DB_Manager.java
@@ -8,16 +8,17 @@ import android.database.sqlite.SQLiteOpenHelper;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.Cursor;
 
+import java.io.Serializable;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class DB_Manager extends SQLiteOpenHelper {
+public class DB_Manager extends SQLiteOpenHelper{
 
     private final static String DB_NAME = "DB_1";
-    private final static int DB_VER = 2;
+    private final static int DB_VER = 3;
     private Context context;
     ArrayList<Company> companies;
     ArrayList<Customer> customers;
@@ -180,9 +181,8 @@ public class DB_Manager extends SQLiteOpenHelper {
         values.put(COMPANY_PASSWORD, company.getPassword());
 
         long newRowId = db.insert(TBL_COMPANIES, null, values);
-        companies.add(company);
         db.close();
-        if(newRowId != -1){
+        if(newRowId !=-1){
             company.setId((int)newRowId);
             companies.add(company);
         }
@@ -213,6 +213,7 @@ public class DB_Manager extends SQLiteOpenHelper {
     }
 
     public void deleteCompany(int companyID) throws myException {
+        companies = getAllCompanies();
         Company tobeDeleted = null;
         boolean flag = false;
         for (Company c : companies) {
@@ -223,13 +224,14 @@ public class DB_Manager extends SQLiteOpenHelper {
             }
         }
         if (flag) {
-            for(Coupon c : tobeDeleted.getCoupons())
-            {
-                deleteCoupon(c);
+            if(tobeDeleted!=null && tobeDeleted.getCoupons()!=null) {
+                for (Coupon c : tobeDeleted.getCoupons()) {
+                    deleteCoupon(c);
+                }
             }
             companies.remove(tobeDeleted); // deleting it from the arraylist because all object in the list are references
             SQLiteDatabase db = getWritableDatabase();
-            db.delete(TBL_COMPANIES, COMPANY_ID + "= ?", new String[]{(Integer.toString(companyID))});
+            db.delete(TBL_COMPANIES, COMPANY_ID + "= ?", new String[]{(companyID+"")});
 
         } else throw new myException("Delete failed No company found with id =" + companyID + "!");
 
@@ -237,6 +239,7 @@ public class DB_Manager extends SQLiteOpenHelper {
 
     public ArrayList<Company> getAllCompanies() {
         String[] fields = {COMPANY_ID, COMPANY_NAME, COMPANY_EMAIL, COMPANY_PASSWORD};
+        ArrayList<Company> companiesCpy = new ArrayList<>();
         String id, name, email, password;
         try {
             Cursor cr = getCursor(TBL_COMPANIES, fields, null);
@@ -246,15 +249,23 @@ public class DB_Manager extends SQLiteOpenHelper {
                     name = cr.getString(1);
                     email = cr.getString(2);
                     password = cr.getString(3);
-                    companies.add(new Company(name, email, password));
+                    companiesCpy.add(new Company(Integer.parseInt(id),name, email, password,null));
                 } while (cr.moveToNext());
-            return companies;
+            return companiesCpy;
         } catch (Exception e) {
             throw e;
         }
     }
 
 
+    public Company getOneCompanyByName(String companyName) {
+        companies = getAllCompanies();
+        for (Company c : companies) {
+            if (c.getName().equals(companyName))
+                return c;
+        }
+        return null;
+    }
     public Company getOneCompany(int CompanyID) {
         companies = getAllCompanies();
         for (Company c : companies) {
@@ -283,22 +294,26 @@ public class DB_Manager extends SQLiteOpenHelper {
                 if(c.getPassword().equals(password))
                     return c.getId();
         }
-        return -1;    }
+        return -1;
+    }
 
 
 
     public void addCustomer(Customer customer) {
         SQLiteDatabase db = getWritableDatabase();
         ContentValues values = new ContentValues();
-        values.put(CUSTOMER_ID, customer.getId());
+        //values.put(CUSTOMER_ID, customer.getId());
         values.put(CUSTOMER_FNAME, customer.getFirstName());
         values.put(CUSTOMER_LNAME, customer.getLastName());
         values.put(CUSTOMER_EMAIL, customer.getEmail());
         values.put(CUSTOMER_PASSWORD, customer.getPassword());
 
-        db.insert(TBL_CUSTOMERS, null, values);
+        long newRowId = db.insert(TBL_CUSTOMERS, null, values);
         db.close();
-        customers.add(customer);
+        if(newRowId !=-1){
+            customer.setId((int)newRowId);
+            customers.add(customer);
+        }
     }
 
 
@@ -362,7 +377,7 @@ public class DB_Manager extends SQLiteOpenHelper {
                     lname = cr.getString(2);
                     email = cr.getString(3);
                     password = cr.getString(4);
-                    customers.add(new Customer(fname, lname, email, password));
+                    customers.add(new Customer(Integer.parseInt(id),fname, lname, email, password));
                 } while (cr.moveToNext());
             return customers;
         } catch (Exception e) {
